process: match Node's exit code and default unhandled-rejection semantics
Store process.exitCode as an int32 instead of a masked u8 so the getter
and the beforeExit/exit event arguments keep the raw value (-1, 300),
reset it on a nullish assignment (including process.exit(null) and
process.exit(undefined)), and re-read it after the exit listeners run
so a listener can override the code passed to process.exit(). Only the
final exit(2) masks to 0-255.

Route the default unhandled-rejection mode through the same path as
--unhandled-rejections=throw, which is Node's default: with no
unhandledRejection listener the rejection is raised as an uncaught
exception so uncaughtExceptionMonitor and uncaughtException listeners
run, and an uncaughtException listener keeps the process alive. This
also removes a duplicate error report on the throw-mode fatal path.

events: read EventEmitter.prototype[kCapture] at emit time instead of
baking the choice in at construction, so the global
EventEmitter.captureRejections setting applies to instances whose
constructor never ran (util.inherits), and duck-type .then on any
non-nullish listener return value like Node so plain thenables are
captured and a throwing `then` getter is routed to the error event.

+++ src/jsc/web_worker.rs
@@ -1250,7 +1250,7 @@ impl WebWorker {
                 // is step 3 below).
                 rare.close_all_socket_groups(unsafe { &*vm_ptr });
             }
-            exit_code = i32::from(vm.exit_handler.exit_code);
+            exit_code = vm.exit_handler.exit_code;
             global_object = Some(vm.global);
         }
